# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for BrandishBot game engine - inventory, crafting, economy, and stats management",
        "title": "BrandishBot API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/osse101/BrandishBot_Go"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/recipes": {
            "get": {
                "description": "Get recipe information. Can filter by item or get all unlocked recipes for a user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafting"
                ],
                "summary": "Get recipes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item name to get recipe for",
                        "name": "item",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username to get unlocked recipes for",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipes or single recipe",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/item/upgrade": {
            "post": {
                "description": "Upgrade an item to a higher tier\nUpgrade an item to a higher tier",
                "consumes": [
                    "application/json",
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "crafting",
                    "crafting"
                ],
                "summary": "Upgrade item",
                "parameters": [
                    {
                        "description": "Upgrade details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpgradeItemRequest"
                        }
                    },
                    {
                        "description": "Upgrade details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpgradeItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpgradeItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Feature locked",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpgradeItemRequest": {
            "type": "object",
            "required": [
                "item",
                "username"
            ],
            "properties": {
                "item": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "handler.UpgradeItemResponse": {
            "type": "object",
            "properties": {
                "new_item": {
                    "type": "string"
                },
                "quantity_upgraded": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}